+++ src/Application/VortexLoader/Converters.Obstruction.cs
@@ -30,11 +30,11 @@ namespace S100Framework.Applications
 
             instance.featureName = ImporterNIS.GetFeatureName(current.OBJNAM, current.NOBJNM);
 
-            if (current.HEIGHT.HasValue && current.HEIGHT.Value != -32767m) {
+            if (current.HEIGHT.HasValue && current.HEIGHT.Value != -32767d) {
                 instance.height = current.HEIGHT.Value;
             }
             else {
-                instance.height = default(decimal?);
+                instance.height = default(double?);
             }
 
             // DODO: Interoperability identifier
@@ -72,18 +72,18 @@ namespace S100Framework.Applications
             }
 
 
-            if (current.VALSOU.HasValue && current.VALSOU.Value != -32767m) {
+            if (current.VALSOU.HasValue && current.VALSOU.Value != -32767d) {
                 instance.valueOfSounding = current.VALSOU.Value;
             }
             else {
-                instance.valueOfSounding = default(decimal?);
+                instance.valueOfSounding = default(double?);
             }
 
             if (current.VERLEN.HasValue) {
                 instance.verticalLength = current.VERLEN.Value;
             }
-            else if (current.VERLEN.HasValue && current.VERLEN.Value == -32767m) {
-                instance.verticalLength = default(decimal?);
+            else if (current.VERLEN.HasValue && current.VERLEN.Value == -32767d) {
+                instance.verticalLength = default(double?);
             }
 
             if (current.WATLEV.HasValue) {
@@ -134,11 +134,11 @@ namespace S100Framework.Applications
 
             instance.featureName = ImporterNIS.GetFeatureName(current.OBJNAM, current.NOBJNM);
 
-            if (current.HEIGHT.HasValue && current.HEIGHT.Value != -32767m) {
+            if (current.HEIGHT.HasValue && current.HEIGHT.Value != -32767d) {
                 instance.height = current.HEIGHT.Value;
             }
             else {
-                instance.height = default(decimal?);
+                instance.height = default(double?);
             }
 
             // DODO: Interoperability identifier
@@ -176,18 +176,18 @@ namespace S100Framework.Applications
             }
 
 
-            if (current.VALSOU.HasValue && current.VALSOU.Value != -32767m) {
+            if (current.VALSOU.HasValue && current.VALSOU.Value != -32767d) {
                 instance.valueOfSounding = current.VALSOU.Value;
             }
             else {
-                instance.valueOfSounding = default(decimal?);
+                instance.valueOfSounding = default(double?);
             }
 
             if (current.VERLEN.HasValue) {
                 instance.verticalLength = current.VERLEN.Value;
             }
-            else if (current.VERLEN.HasValue && current.VERLEN.Value == -32767m) {
-                instance.verticalLength = default(decimal?);
+            else if (current.VERLEN.HasValue && current.VERLEN.Value == -32767d) {
+                instance.verticalLength = default(double?);
             }
 
             if (current.WATLEV.HasValue) {
@@ -238,11 +238,11 @@ namespace S100Framework.Applications
 
             instance.featureName = ImporterNIS.GetFeatureName(current.OBJNAM, current.NOBJNM);
 
-            if (current.HEIGHT.HasValue && current.HEIGHT.Value != -32767m) {
+            if (current.HEIGHT.HasValue && current.HEIGHT.Value != -32767d) {
                 instance.height = current.HEIGHT.Value;
             }
             else {
-                instance.height = default(decimal?);
+                instance.height = default(double?);
             }
 
             // DODO: Interoperability identifier
@@ -280,18 +280,18 @@ namespace S100Framework.Applications
             }
 
 
-            if (current.VALSOU.HasValue && current.VALSOU.Value != -32767m) {
+            if (current.VALSOU.HasValue && current.VALSOU.Value != -32767d) {
                 instance.valueOfSounding = current.VALSOU.Value;
             }
             else {
-                instance.valueOfSounding = default(decimal?);
+                instance.valueOfSounding = default(double?);
             }
 
             if (current.VERLEN.HasValue) {
                 instance.verticalLength = current.VERLEN.Value;
             }
-            else if (current.VERLEN.HasValue && current.VERLEN.Value == -32767m) {
-                instance.verticalLength = default(decimal?);
+            else if (current.VERLEN.HasValue && current.VERLEN.Value == -32767d) {
+                instance.verticalLength = default(double?);
             }
 
             if (current.WATLEV.HasValue) {
