+++ src/Application/VortexLoader/S57_DangersL.cs
@@ -86,11 +86,11 @@ namespace S100Framework.Applications
                                 instance.status = GetStatus(current.STATUS);
                             }
 
-                            if (current.VERLEN.HasValue && current.VERLEN.Value != -32767m) {
+                            if (current.VERLEN.HasValue && current.VERLEN.Value != -32767d) {
                                 instance.verticalLength = current.VERLEN.Value;
                             }
                             else {
-                                instance.verticalLength = default(decimal?);
+                                instance.verticalLength = default(double?);
                             }
 
                             if (current.PLTS_COMP_SCALE.HasValue && current.SHAPE != null) {
@@ -159,11 +159,11 @@ namespace S100Framework.Applications
                                 }
 
 
-                                if (current.VALSOU.HasValue && current.VALSOU.Value != -32767m) {
+                                if (current.VALSOU.HasValue && current.VALSOU.Value != -32767d) {
                                     instance.valueOfSounding = current.VALSOU.Value;
                                 }
                                 else {
-                                    instance.valueOfSounding = default(decimal?);
+                                    instance.valueOfSounding = default(double?);
                                 }
 
                                 if (current.SOUACC.HasValue) {
@@ -223,11 +223,11 @@ namespace S100Framework.Applications
                            //     }
 
                            //     instance.featureName = GetFeatureName(current.OBJNAM, current.NOBJNM);                            
-                           //if (current.HEIGHT.HasValue && current.HEIGHT.Value != -32767m) {
+                           //if (current.HEIGHT.HasValue && current.HEIGHT.Value != -32767d) {
                            //     instance.height = current.HEIGHT.Value;
                            // }
                            // else {
-                           //     instance.height = default(decimal?);
+                           //     instance.height = default(double?);
                            // }
 
                            //     // DODO: Interoperability identifier
@@ -263,11 +263,11 @@ namespace S100Framework.Applications
                            //     }
 
 
-                           //     if (current.VALSOU.HasValue && current.VALSOU.Value != -32767m) {
+                           //     if (current.VALSOU.HasValue && current.VALSOU.Value != -32767d) {
                            //         instance.valueOfSounding = current.VALSOU.Value;
                            //     }
                            //     else {
-                           //         instance.valueOfSounding = default(decimal?);
+                           //         instance.valueOfSounding = default(double?);
                            //     }
 
 
@@ -275,8 +275,8 @@ namespace S100Framework.Applications
                            //     if (current.VERLEN.HasValue) {
                            //         instance.verticalLength = current.VERLEN.Value;
                            //     }
-                           //     else if (current.VERLEN.HasValue && current.VERLEN.Value == -32767m) {
-                           //         instance.verticalLength = default(decimal?);
+                           //     else if (current.VERLEN.HasValue && current.VERLEN.Value == -32767d) {
+                           //         instance.verticalLength = default(double?);
                            //     }
